requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.2
blinker==1.6.2
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
distro==1.8.0
dnspython==2.4.2
dominate==2.8.0
email-validator==2.1.0.post1
et_xmlfile==1.1.0
filelock==3.12.4
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Reuploaded==1.3.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
fsspec==2023.10.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.22.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
networkx==3.2.1
numpy==1.24.4
openai==1.3.5
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
pydantic==2.5.3
pydantic_core==2.14.6
pyodbc==5.0.1
PyYAML==6.0.1
regex==2023.12.25
reportlab==4.0.8
safetensors==0.4.2
setuptools==68.2.2
sniffio==1.3.0
SQLAlchemy==2.0.23
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.39.3
typing-inspection==0.4.0
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==2.3.8
WTForms==3.1.1
